Hook up compact and v2 to get most things working
This is the big drop to get all the major features working
properly atop the v2 and compact index APIs rather than the now-
defunct v1 dependencies API.

I was guessing my way through this codebase a bit but the new APIs
are hooked up as follows:

* The v1 dependencies API (dependencies?gems=foo,bar) has been
  mostly replaced with the compact index API. This API provides
  all versions of one specified gem. Any features that depend on
  the dependencies API to get multiple gems at once are still
  broken and will need to iterate over all given gems using the
  compact index.
* The creation of gemspecs and poms from the old gzipped spec
  format has been moved to the v2 gem info API.

All tests are passing in mavengem-protocol, which only depends on
features like poms and gems, all of which have been moved to the
new APIs.

This is close to addressing #8.