Update dependencies to build a snapshot
Group ID is changed to "com.headius" because I do not have push
privileges for "org.torquebox.mojo".

Inter-library dependencies switched to use "com.headius" for
snapshot deploy, but build still uses the previous release of
mavengem extension (did not seem to like me using snapshot
extensions from the same project, or at least refused to resolve
them.

maven-tools updated to latest to fix a bug in gem logic that tried
to gsub! on an empty string that now appears to be coming in
frozen:

https://github.com/torquebox/maven-tools/commit/11887434e2838ee326e770259c0563749573414e#diff-e4394efc5b28617f42311627dd8af1193c01d1a39b939959085c6836f3e9e5b7R149

With these changes, I have successfully been able to:

* Deploy 1.0.4-SNAPSHOT versions of all three artifacts
* Show JRuby build failing with a patched 1.0.3 that uses a
  non-"Java" User-Agent when contacting rubygems.org (causing the
  v1 deps API to reject requests and gems fail to resolve.
* Show JRuby build working with the same User-Agent patch but the
  new logic to use v2 and compact APIs.

Nearly complete solution for #8 but not all features in
rubygems-tools are updated and not all v1 deps API and gz data
file uses have been purged yet.