Tweaks for specs to run with new API
Mostly these needed to be updated to use real gems, since I did
not want to continue to fake out the made-up gems, and to expect
the different caching and behavior of the parts of the API hooked
up to the v2 or compact index APIs.

There's some form-fitting here, and a bunch of specs are testing
defunct APIs using local data, but this gets all unit tests
passing.